Move TraitsEditor to shared/components
