+++ web/packages/teleport/src/Users/UserAddEdit/UserAddEdit.tsx
@@ -30,7 +30,7 @@ import Validation from 'shared/components/Validation';
 import { requiredField } from 'shared/components/Validation/rules';
 
 import UserTokenLink from './../UserTokenLink';
-import { TraitsEditor } from './TraitsEditor';
+import { TraitsEditor } from 'shared/components/TraitsEditor';
 import useDialog, { Props } from './useDialog';
 
 export default function Container(props: Props) {
