--- web/packages/teleport/src/Users/UserAddEdit/useDialog.test.ts
@@ -1,35 +1,0 @@
-/**
- * Teleport
- * Copyright (C) 2023  Gravitational, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-import { traitsToTraitsOption } from './useDialog';
-
-describe('Test traitsToTraitsOption', () => {
-  test.each`
-    name                                         | trait                | expected
-    ${'trait with values (valid)'}               | ${{ t: ['a', 'b'] }} | ${[{ traitKey: { label: 't', value: 't' }, traitValues: [{ label: 'a', value: 'a' }, { label: 'b', value: 'b' }] }]}
-    ${'empty trait (invalid)'}                   | ${{ t: [] }}         | ${[]}
-    ${'trait with empty string (invalid)'}       | ${{ t: [''] }}       | ${[]}
-    ${'trait with null value (invalid)'}         | ${{ t: null }}       | ${[]}
-    ${'trait with null array (invalid)'}         | ${{ t: [null] }}     | ${[]}
-    ${'trait with first empty string (invalid)'} | ${{ t: ['', 'a'] }}  | ${[{ traitKey: { label: 't', value: 't' }, traitValues: [{ label: '', value: '' }, { label: 'a', value: 'a' }] }]}
-    ${'trait with last empty element (valid)'}   | ${{ t: ['a', ''] }}  | ${[{ traitKey: { label: 't', value: 't' }, traitValues: [{ label: 'a', value: 'a' }, { label: '', value: '' }] }]}
-  `('$name', ({ trait, expected }) => {
-    const result = traitsToTraitsOption(trait);
-    expect(result).toEqual(expected);
-  });
-});
